# Fix escaping keys in MarshalJSON