This is a fix for https://github.com/iancoleman/orderedmap/issues/17#issuecomment-757491931, the issue about MarshalJSON. Also improves the performance by reducing the allocations. ~This change follows #19.~

<details>
<summary>BenchmarkUnmarshalJSON</summary>

```go
func BenchmarkMarshalJSON(b *testing.B) {
	s := `{
  "number": 4,
  "string": "x",
  "z": 1,
  "a": "should not break with unclosed { character in value",
  "b": 3,
  "slice": [
    "1",
    1
  ],
  "orderedmap": {
    "e": 1,
    "a { nested key with brace": "with a }}}} }} {{{ brace value",
	"after": {
		"link": "test {{{ with even deeper nested braces }"
	}
  },
  "test\"ing": 9,
  "after": 1,
  "multitype_array": [
    "test",
	1,
	{ "map": "obj", "it" : 5, ":colon in key": "colon: in value" },
	[{"inner": "map"}]
  ],
  "should not break with { character in key": 1
}`
	o := New()
	json.Unmarshal([]byte(s), &o)
	for i := 0; i < b.N; i++ {
		json.Marshal(o)
	}
}
```
</details>

```
benchmark                   old ns/op     new ns/op     delta
BenchmarkMarshalJSON-16     15687         14535         -7.34%

benchmark                   old allocs     new allocs     delta
BenchmarkMarshalJSON-16     94             33             -64.89%

benchmark                   old bytes     new bytes     delta
BenchmarkMarshalJSON-16     9309          2195          -76.42%
```